Overall transactions chart

+++ backend/src/controllers/incomeControllers.ts
@@ -3,8 +3,11 @@ import Income, { IIncome } from "../models/incomeModel";
 
 // Get all incomes - api/incomes
 const getIncomes = async (req: Request, res: Response, next: NextFunction) => {
+  const { _id } = req.user;
   try {
-    const incomes: IIncome[] = await Income.find();
+    const incomes: IIncome[] = await Income.find({
+      user: _id,
+    });
     res.status(200).json(incomes);
   } catch (error) {
     next(error);
@@ -17,13 +20,15 @@ const createIncome = async (
   res: Response,
   next: NextFunction
 ) => {
+  const { _id } = req.user;
   const { name, amount, date, source, description } = req.body;
-  if (!name || !amount || !date || !source) {
+  if (!name || !amount || !source) {
     res.status(400);
     throw new Error("Please enter required fields");
   }
   try {
     const newIncome = await Income.create({
+      user: _id,
       name: name,
       amount: amount,
       date: date,
@@ -38,12 +43,17 @@ const createIncome = async (
 
 // Edit income - api/incomes/:id
 const editIncome = async (req: Request, res: Response, next: NextFunction) => {
+  const { _id } = req.user;
   const { id } = req.params;
   try {
     const income = await Income.findById(id);
     if (!income) {
       res.status(404);
       throw new Error("Income not found");
+    }
+    if (income.user.toString() !== _id.toString()) {
+      res.status(403);
+      throw new Error("Not authorized");
     }
     const updatedIncome = await Income.findByIdAndUpdate(id, req.body, {
       new: true,
@@ -60,12 +70,17 @@ const deleteIncome = async (
   res: Response,
   next: NextFunction
 ) => {
+  const { _id } = req.user;
   const { id } = req.params;
   try {
     const income = await Income.findById(id);
     if (!income) {
       res.status(404);
       throw new Error("Income not found");
+    }
+    if (income.user.toString() !== _id.toString()) {
+      res.status(403);
+      throw new Error("Not authorized");
     }
     const deletedIncome = await Income.findByIdAndDelete(id);
     res.status(200).json(deletedIncome?._id);
